Updated build process
