--- buildSrc/GroupMappings.ts
@@ -1,26 +1,0 @@
-// @ts-ignore
-import {StringDictonary} from "./BuildThemes";
-
-const GroupToNameMapping: StringDictonary<string> = {
-  "Kill la Kill": "KillLaKill: ",
-  "Re Zero": "Re:Zero: ",
-  "Fate": "Fate: ",
-  "Gate": "Gate: ",
-  "Steins Gate": "Steins Gate: ",
-  "Literature Club": "DDLC: ",
-  "KonoSuba": "KonoSuba: ",
-  "Love Live": "Love Live: ",
-  "OreGairu": "OreGairu: ",
-  "Darling in the Franxx": "Franxx: ",
-  "Bunny Senpai": "BunnySenpai: ",
-  "Miss Kobayashi's Dragon Maid": "DM: ",
-  "Mistress": "Miss: ",
-  "DanganRonpa": "DR: ",
-  "Evangelion": "EVA: ",
-  "Miscellaneous": "Misc: ",
-  "High School DxD" : "DxD: ",
-  "Sword Art Online" : "SAO: ",
-  "Lucky Star" : "LS: ",
-};
-
-export default GroupToNameMapping;
