+++ buildSrc/assets/templates/base.colors.template.json
