+++ buildSrc/assets/templates/base.laf.template.json
