+++ buildSrc/assets/templates/base.syntax.template.json
