+++ buildSrc/assets/templates/dark.base.laf.template.json
