+++ buildSrc/assets/templates/dark.colors.template.json
