+++ buildSrc/assets/templates/light.colors.template.json
