+++ buildSrc/assets/themes/bunnySenpai/mai/light/mai.light.vsCode.definition.json
