+++ buildSrc/assets/themes/fate/ishtar/dark/ishtar.dark.vsCode.definition.json
