+++ buildSrc/assets/themes/fate/ishtar/light/ishtar.light.vsCode.definition.json
