+++ buildSrc/assets/themes/fate/rin/dark/rin.dark.vsCode.definition.json
