+++ buildSrc/assets/themes/franxx/zeroTwo/dark/zero.two.dark.vsCode.definition.json
