+++ buildSrc/assets/themes/franxx/zeroTwo/light/zero.two.light.vsCode.definition.json
