+++ buildSrc/assets/themes/gate/rory/dark/rory.dark.vsCode.definition.json
