+++ buildSrc/assets/themes/literature/monika/dark/monika.dark.vsCode.definition.json
