+++ buildSrc/assets/themes/literature/monika/light/monika.light.vsCode.definition.json
