+++ buildSrc/assets/themes/literature/sayori/dark/sayori.dark.vsCode.definition.json
