+++ buildSrc/assets/themes/literature/sayori/light/sayori.light.vsCode.definition.json
