+++ buildSrc/assets/themes/literature/yuri/dark/yuri.dark.vsCode.definition.json
