+++ buildSrc/assets/themes/literature/yuri/light/yuri.light.vsCode.definition.json
