+++ buildSrc/assets/themes/loveLive/umi/dark/umi.dark.vsCode.definition.json
