+++ buildSrc/assets/themes/reZero/beatrice/beatrice.vsCode.definition.json
