+++ buildSrc/assets/themes/steinsGate/kurisu/dark/kurisu.dark.vsCode.definition.json
